ml/torch_explore/requirements.txt
tensorflow-gpu~=2.5.3

numpy~=1.19.5
torch~=1.10.2+cu113
Pillow~=7.0.0